requirements.txt
Django
south
johnny-cache
django-debug
django-taggit
BeautifulSoup
django-form-utils